photos getting better, infinite loop but y the fuck ???

+++ d02/photos.php
@@ -4,12 +4,20 @@
 function get_imgurl($tab)
 {
 	$i = 0;
-	while ($tab[$i] != NULL)
+	while ($tab[0][$i] != NULL)
 	{
-		$len = strlen($tab[$i]) - 11;
-		$ret[$i] = substr($tab[$i++], 10, $len);
+		echo ("\$tab[0][".$i.'] = '.$tab[0][$i]."\n");
+		preg_match('#src=\"([A-Za-z0-9\.:/\_?%=-]+)#', $tab[0][$i], $ret[$i]);
+		$i++;
+		//$len = strlen($tab[$i]) - 11;
+		//$ret[$i] = substr($tab[$i++], 10, $len);
 	}
-	return $ret;
+	echo "\n------ RET -----\n";
+	print_r($ret);
+	$i = 0;
+	while ($ret[$i])
+		$return[$i] = $ret[$i][1];
+	return $return;
 }
 
 function get_images($argv)
@@ -18,7 +26,7 @@ function get_images($argv)
 	curl_setopt($c, CURLOPT_RETURNTRANSFER, 1);
 	$cont = curl_exec($c);
 	curl_close($c);
-	preg_match('#<img src=[\"\'][\w-:/.]*[\"\']#', $cont, $tab);
+	preg_match_all('#<img (.)+>#', $cont, $tab);
 	$imgurl = get_imgurl($tab);
 	return $imgurl;
 }
@@ -60,10 +68,12 @@ if ($argv[1])
 	$file = substr($argv[1], 7);
 	$file = "./".$file."/";
 	$imgurls = get_images($argv);
-	if (is_dir($file))
-		rmdir($file);
-	if (!mkdir($file))
-		echo "MKDIR: ERROR";
-	dl_images($imgurls, $file);
+	if ($imgurls) {
+		if (is_dir($file))
+			rmdir($file);
+		if (!mkdir($file))
+			echo "MKDIR: ERROR";
+		dl_images($imgurls, $file);
+	}
 }
 ?>
